+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add new frame fields for MetricKit flamegraphs. ([#5539](https://github.com/getsentry/relay/pull/5539))
 - Apply clock drift correction to logs and trace metrics. ([#5609](https://github.com/getsentry/relay/pull/5609))
 - Add Culture context to event schema. ([#5615](https://github.com/getsentry/relay/pull/5615))
+- Trim spans with a new EAP trimming processor. ([#5616](https://github.com/getsentry/relay/pull/5616))
 
 **Internal**: